Debugging and CLeanpup
