--- catkin_ws/src/controls/scripts/drc_math.py
@@ -1,94 +1,0 @@
-#!/usr/bin/env python
-
-import numpy as np
-from tf.transformations import unit_vector, quaternion_multiply, quaternion_conjugate
-
-def quat_vec_mult(q1, v1):
-    """Rotate vector v1 by quaternion q1, and return the resulting vector.
-    From https://answers.ros.org/question/196149/how-to-rotate-vector-by-quaternion-in-python/
-    """
-    q2 = list(v1)
-    q2.append(0.0)
-    return quaternion_multiply(
-        quaternion_multiply(q1, q2), 
-        quaternion_conjugate(q1)
-    )[:3]
-
-
-##### Following functions from https://github.com/NxRLab/ModernRobotics/ #####
-
-def VecToso3(omg):
-    """Converts a 3-vector to an so(3) representation
-    :param omg: A 3-vector
-    :return: The skew symmetric representation of omg
-    Example Input:
-        omg = np.array([1, 2, 3])
-    Output:
-        np.array([[ 0, -3,  2],
-                  [ 3,  0, -1],
-                  [-2,  1,  0]])
-    """
-    return np.array([[0,      -omg[2],  omg[1]],
-                     [omg[2],       0, -omg[0]],
-                     [-omg[1], omg[0],       0]])
-
-def RpToTrans(R, p):
-    """Converts a rotation matrix and a position vector into homogeneous
-    transformation matrix
-    :param R: A 3x3 rotation matrix
-    :param p: A 3-vector
-    :return: A homogeneous transformation matrix corresponding to the inputs
-    Example Input:
-        R = np.array([[1, 0,  0],
-                      [0, 0, -1],
-                      [0, 1,  0]])
-        p = np.array([1, 2, 5])
-    Output:
-        np.array([[1, 0,  0, 1],
-                  [0, 0, -1, 2],
-                  [0, 1,  0, 5],
-                  [0, 0,  0, 1]])
-    """
-    return np.r_[np.c_[R, p], [[0, 0, 0, 1]]]
-
-def TransToRp(T):
-    """Converts a homogeneous transformation matrix into a rotation matrix
-    and position vector
-    :param T: A homogeneous transformation matrix
-    :return R: The corresponding rotation matrix,
-    :return p: The corresponding position vector.
-    Example Input:
-        T = np.array([[1, 0,  0, 0],
-                      [0, 0, -1, 0],
-                      [0, 1,  0, 3],
-                      [0, 0,  0, 1]])
-    Output:
-        (np.array([[1, 0,  0],
-                   [0, 0, -1],
-                   [0, 1,  0]]),
-         np.array([0, 0, 3]))
-    """
-    T = np.array(T)
-    return T[0: 3, 0: 3], T[0: 3, 3]
-
-def Adjoint(T):
-    """Computes the adjoint representation of a homogeneous transformation
-    matrix
-    :param T: A homogeneous transformation matrix
-    :return: The 6x6 adjoint representation [AdT] of T
-    Example Input:
-        T = np.array([[1, 0,  0, 0],
-                      [0, 0, -1, 0],
-                      [0, 1,  0, 3],
-                      [0, 0,  0, 1]])
-    Output:
-        np.array([[1, 0,  0, 0, 0,  0],
-                  [0, 0, -1, 0, 0,  0],
-                  [0, 1,  0, 0, 0,  0],
-                  [0, 0,  3, 1, 0,  0],
-                  [3, 0,  0, 0, 0, -1],
-                  [0, 0,  0, 0, 1,  0]])
-    """
-    R, p = TransToRp(T)
-    return np.r_[np.c_[R, np.zeros((3, 3))],
-                 np.c_[np.dot(VecToso3(p), R), R]]
